# src/main/resources/openapi/openapi.yml
openapi: 3.0.3
info:
  title: Library management system APIs
  contact:
    email: romachervinko@gmail.com
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /book-state:
    get:
      tags:
        - book-state
      summary: Get book states
      operationId: getBookStates
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BookStateResponseDTO'
                  meta:
                    type: object
    post:
      tags:
        - book-state
      summary: Create book state
      operationId: createBookState
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/BookStateCreateRequestDTO'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookStateResponseDTO'
      x-codegen-request-body-name: body
    put:
      tags:
        - book-state
      summary: Updated book state
      operationId: updateBookState
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/BookStateUpdateRequestDTO'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookStateResponseDTO'
  /book-state/{id}:
    get:
      tags:
        - book-state
      summary: Get publishinghouse by id
      operationId: getBookStateById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookStateResponseDTO'
    delete:
      tags:
        - book-state
      summary: Deletes book state
      operationId: deleteBookStateById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '404':
          description: Theme not found
          content: { }
  /book-state/name/{name}:
    get:
      tags:
        - book-state
      summary: Get publishinghouse by name
      operationId: getBookStateByName
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookStateResponseDTO'
components:
  schemas:
...
    BookStateCreateRequestDTO:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
      required:
        - name
    BookStateUpdateRequestDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
          minLength: 1
        description:
          type: string
      required:
        - id
        - name
    BookStateResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string